pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
faiss-cpu = "^1.7.4"
pandas = "^2.1.3"
tqdm = "^4.66.1"
gradio = "^4.7.1"
openai = "^1.3.6"
flask = "^3.0.0"
pymupdf = "1.23.7"
fitz = "^0.0.1.dev2"
langchain = "^0.1.14"
pinecone-client = "^3.2.2"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"